Small changes in bus.rs and bus8080

+++ kernel/src/hil/bus8080.rs
@@ -24,7 +24,7 @@ impl BusWidth {
 
 pub trait Bus8080<'a> {
     /// Set the address to write to
-    fn set_addr(&self, addr_width: BusWidth, addr: usize) -> Result<(), ErrorCode>;
+    fn set_addr(&self, addr_width: BusWidth, addr: u16) -> Result<(), ErrorCode>;
 
     /// Write data items to the previously set address
     fn write(
